+++ ui/app/tsconfig.json
@@ -34,7 +34,7 @@
     "src/**/*.vue",
     "tests/**/*.ts",
     "tests/**/*.tsx"
-  ],
+ ],
   "exclude": [
     "node_modules"
   ]
