+++ ui/core/src/hooks/poolCalculator.ts
@@ -1,29 +1,7 @@
 import { computed } from "@vue/reactivity";
 import { Ref } from "vue";
-import { Asset, AssetAmount, IAssetAmount, Pair } from "../../../../core";
-
-function buildAsset(val: string | null) {
-  return val === null ? val : Asset.get(val);
-}
-
-function buildAssetAmount(asset: Asset | null, amount: string) {
-  return asset ? AssetAmount(asset, amount) : asset;
-}
-
-export function useField(amount: Ref<string>, symbol: Ref<string | null>) {
-  const asset = computed(() => {
-    return buildAsset(symbol.value);
-  });
-
-  const fieldAmount = computed(() => {
-    return buildAssetAmount(asset.value, amount.value);
-  });
-
-  return {
-    fieldAmount,
-    asset,
-  };
-}
+import { IAssetAmount, Pair } from "../entities";
+import { useField } from "./useField";
 
 export function usePoolCalculator(input: {
   fromAmount: Ref<string>;
