+++ ui/core/src/hooks/swapCalculator.ts
@@ -1,29 +1,6 @@
-import { computed, effect } from "@vue/reactivity";
-import { Ref } from "vue";
-import { Asset, AssetAmount, IAssetAmount, Pair } from "../../../../core";
-
-function buildAsset(val: string | null) {
-  return val === null ? val : Asset.get(val);
-}
-
-function buildAssetAmount(asset: Asset | null, amount: string) {
-  return asset ? AssetAmount(asset, amount) : asset;
-}
-
-export function useField(amount: Ref<string>, symbol: Ref<string | null>) {
-  const asset = computed(() => {
-    return buildAsset(symbol.value);
-  });
-
-  const fieldAmount = computed(() => {
-    return buildAssetAmount(asset.value, amount.value);
-  });
-
-  return {
-    fieldAmount,
-    asset,
-  };
-}
+import { Ref, computed, effect } from "@vue/reactivity";
+import { Asset, AssetAmount, IAssetAmount, Pair } from "../entities";
+import { useField } from "./useField";
 
 function useBalances(balances: Ref<AssetAmount[]>) {
   return computed(() => {
