+++ ui/core/src/index.ts
@@ -2,3 +2,4 @@ export * from "./entities";
 export * from "./api";
 export * from "./actions";
 export * from "./store";
+export * from "./hooks";
